% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRefresh.R
\name{koRefresh}
\alias{koRefresh}
\alias{koAutoRefresh}
\title{Refresh Komodo interface (active object lists and R object browser)}
\usage{
koRefresh(force = FALSE)

koAutoRefresh(...)
}
\arguments{
\item{force}{do we force refresh, even if data have not changed?}

\item{...}{any argument (ignored, but useful for \code{addTaskCallback()}).}
}
\value{
Both functions return \code{TRUE} on success.
}
\description{
These function manage to refresh the list of active objects in Komodo and
also the data displayed in the Komodo R object browser. They should not be
called directly by the end-user.
}
\seealso{
\code{\link[=koCmd]{koCmd()}}
}
\concept{interprocess communication Komodo}
\keyword{misc}
