\name{svKomodo-package}
\alias{svKomodo-package}
\alias{svKomodo}
\docType{package}

\title{ SciViews GUI API - Functions to manage communication with Komodo Edit/IDE }
\description{
  The SciViews svKomodo package provides function to communicate with SciViews
  Komodo, which is Komodo Edit or IDE with the SciViews-K plugin installed
  (see http://www.sciviews.org/SciViews-K).
}
\details{
  \tabular{ll}{
    Package: \tab svKomodo\cr
    Type: \tab Package\cr
    Version: \tab 0.9-60\cr
    Date: \tab 2013-02-11\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API communication with Komodo Edit/IDE }
